requirements.txt
torch
torchvision
torchaudio
transformers
datasets
transformers
datasets
evaluate
peft
accelerate
bert_score
rouge_score
nltk
wandb
pyyaml
python-dotenv
sentencepiece
protobuf